refactor: move hooks to its own directory
