+++ src/hooks/useAppState.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { AppStateContext } from "./AppStateContext";
+import { AppStateContext } from "../AppStateContext";
 
 export const useAppState = () => {
   return useContext(AppStateContext);
